requirements.txt
flask
Flask-SQLAlchemy
flask-login
PyJWT
requests
docker

